collapsed a type hierarchy and made tests work on my laptop
