+++ src/test/Bau.NuGet.Test.Unit/Bau.NuGet.Test.Unit.csproj
@@ -57,7 +57,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="NuGetCliCommandRequestBaseFacts.cs" />
-    <Compile Include="NuGetCliDownloadCommandRequestBaseFacts.cs" />
     <Compile Include="NuGetCliLocatorFacts.cs" />
     <Compile Include="NuGetPackFacts.cs" />
     <Compile Include="NuGetPushFacts.cs" />
