+++ src/test/Bau.NuGet.Test.Unit/NuGetRestoreFacts.cs
@@ -16,7 +16,7 @@ namespace BauNuGet.Test.Unit
     using Xunit.Extensions;
 
     public static class NuGetRestoreFacts
-    {        
+    {
         [Fact]
         public static void CanRestorePackagesUsingCli()
         {
@@ -79,6 +79,122 @@ namespace BauNuGet.Test.Unit
             task.Requests.OfType<NuGetCliRestoreCommandRequest>().All(r => r.PackagesDirectory == fakeDirName).Should().BeTrue();
             task.Requests.OfType<NuGetCliRestoreCommandRequest>().Select(x => x.TargetSolutionOrPackagesConfig).Should().Contain("file1");
             task.Requests.OfType<NuGetCliRestoreCommandRequest>().Select(x => x.TargetSolutionOrPackagesConfig).Should().Contain("file2");
+        }
+
+        [Fact]
+        public static void PropertySourceCli()
+        {
+            // arrange
+            var normal = new NuGetCliRestoreCommandRequest();
+            var multiple = new NuGetCliRestoreCommandRequest();
+            multiple.Source.Add(@"http://source1/api");
+            multiple.Source.Add(@"C:\some folder\");
+
+            // act
+            var normalArgs = normal.CreateCommandLineArguments();
+            var multipleArgs = multiple.CreateCommandLineArguments();
+
+            // assert
+            normalArgs.Any(x => x.StartsWith("-Source")).Should().BeFalse();
+            multipleArgs.Should().Contain(@"-Source http://source1/api");
+            multipleArgs.Any(x => x.StartsWith("-Source") && x.Contains(@"C:\some folder")).Should().BeTrue();
+        }
+
+        [Fact]
+        public static void PropertySourceFluent()
+        {
+            // arrange
+            var normal = new NuGetCliRestoreCommandRequest();
+            var multiple = new NuGetCliRestoreCommandRequest();
+
+            // act
+            multiple
+                .WithSource(@"http://source1/api")
+                .WithSource(@"C:\some folder\");
+
+            // assert
+            normal.Source.Should().BeEmpty();
+            multiple.Source.Should().Equal(new[] { @"http://source1/api", @"C:\some folder\" });
+        }
+
+        [Fact]
+        public static void PropertyNoCacheCli()
+        {
+            // arrange
+            var normal = new NuGetCliRestoreCommandRequest();
+            var enabled = new NuGetCliRestoreCommandRequest();
+            enabled.NoCache = true;
+            var disabled = new NuGetCliRestoreCommandRequest();
+            disabled.NoCache = false;
+
+            // act
+            var normalArgs = normal.CreateCommandLineArguments();
+            var enabledArgs = enabled.CreateCommandLineArguments();
+            var disabledArgs = disabled.CreateCommandLineArguments();
+
+            // assert
+            normalArgs.Should().NotContain("-NoCache");
+            enabledArgs.Should().Contain("-NoCache");
+            disabledArgs.Should().NotContain("-NoCache");
+        }
+
+        [Fact]
+        public static void PropertyNoCacheFluent()
+        {
+            // arrange
+            var normal = new NuGetCliRestoreCommandRequest();
+            var enabled = new NuGetCliRestoreCommandRequest();
+            var disabled = new NuGetCliRestoreCommandRequest();
+
+            // act
+            normal.WithNoCache();
+            enabled.WithNoCache(true);
+            disabled.WithNoCache(false);
+
+            // assert
+            normal.NoCache.Should().BeTrue();
+            enabled.NoCache.Should().BeTrue();
+            disabled.NoCache.Should().BeFalse();
+        }
+
+        [Fact]
+        public static void PropertyDisableParallelProcessingCli()
+        {
+            // arrange
+            var normal = new NuGetCliRestoreCommandRequest();
+            var enabled = new NuGetCliRestoreCommandRequest();
+            enabled.DisableParallelProcessing = true;
+            var disabled = new NuGetCliRestoreCommandRequest();
+            disabled.DisableParallelProcessing = false;
+
+            // act
+            var normalArgs = normal.CreateCommandLineArguments();
+            var enabledArgs = enabled.CreateCommandLineArguments();
+            var disabledArgs = disabled.CreateCommandLineArguments();
+
+            // assert
+            normalArgs.Should().NotContain("-DisableParallelProcessing");
+            enabledArgs.Should().Contain("-DisableParallelProcessing");
+            disabledArgs.Should().NotContain("-DisableParallelProcessing");
+        }
+
+        [Fact]
+        public static void PropertyDisableParallelProcessingFluent()
+        {
+            // arrange
+            var normal = new NuGetCliRestoreCommandRequest();
+            var enabled = new NuGetCliRestoreCommandRequest();
+            var disabled = new NuGetCliRestoreCommandRequest();
+
+            // act
+            normal.WithDisableParallelProcessing();
+            enabled.WithDisableParallelProcessing(true);
+            disabled.WithDisableParallelProcessing(false);
+
+            // assert
+            normal.DisableParallelProcessing.Should().BeTrue();
+            enabled.DisableParallelProcessing.Should().BeTrue();
+            disabled.DisableParallelProcessing.Should().BeFalse();
         }
     }
 }
